修改项目结构中

+++ .eslintrc.js
@@ -14,6 +14,6 @@ module.exports = {
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'quotes': ['error', 'double']
+    'quotes': ['error', 'single']
   }
 }
